requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
rembg==2.0.30
Pillow==10.0.0
requests==2.31.0
